Move union into select becasue they are really building the same query
