+++ src/index.ts
@@ -13,7 +13,6 @@ export * from './core/Query.ts'
 export * from './core/query/Delete.ts'
 export * from './core/query/Insert.ts'
 export * from './core/query/Select.ts'
-export * from './core/query/Union.ts'
 export * from './core/query/Update.ts'
 export * from './core/Selection.ts'
 export * from './core/Sql.ts'
