+++ src/postgres.ts
@@ -6,7 +6,7 @@ export {
   intersectAll,
   union,
   unionAll
-} from './core/query/Union.ts'
+} from './core/query/Select.ts'
 export {schema as pgSchema} from './core/Schema.ts'
 export {alias, table as pgTable} from './core/Table.ts'
 export * from './postgres/builder.ts'
